+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added typebot integration activation
 * Now you can register several typebots with triggers
 * Media sent to typebot now goes as a template string, example: imageMessage:MESSAGE_ID
+* Organization configuration and logo in chatwoot bot contact
 
 ### Fixed